+++ src/components/admin/product/ProductInfo.tsx
@@ -1,0 +1,44 @@
+import React from "react";
+import { Product } from "@/lib/admin/products/IProducts";
+import { StockDetailsData } from "@/lib/stock/IStock";
+
+interface ProductInfoProps {
+  product: Product;
+  stockDetails?: StockDetailsData[];
+  isStockLoading?: boolean;
+}
+
+const ProductInfo: React.FC<ProductInfoProps> = ({ 
+  product, 
+  stockDetails, 
+  isStockLoading = false 
+}) => {
+  // Calcular stock total
+  const totalStock = stockDetails 
+    ? stockDetails.reduce((acc, detail) => acc + detail.amount, 0)
+    : 0;
+
+  return (
+    <div className="space-y-2">
+      {[
+        { label: "Código", value: product.code },
+        { label: "Precio", value: `${product.price?.toLocaleString()} Gs` },
+        { label: "Costo", value: `${product.cost?.toLocaleString()} Gs` },
+        { 
+          label: "Stock", 
+          value: isStockLoading 
+            ? "Cargando..." 
+            : totalStock.toString()
+        },
+        { label: "Categoría", value: product.category },
+      ].map(({ label, value }) => (
+        <div key={label} className="flex">
+          <span className="text-gray-600 w-24">{label}:</span>
+          <span className="flex-grow text-right">{value}</span>
+        </div>
+      ))}
+    </div>
+  );
+};
+
+export default ProductInfo;
