+++ src/components/admin/product/ProductList.tsx
@@ -1,12 +1,12 @@
 "use client";
 
-import React, { useCallback, useEffect, useState } from "react";
+import React from "react";
 import { useRouter } from "next/navigation";
 import { Button } from "@/components/ui/button";
-import { getProducts } from "@/lib/admin/products/getProducts";
-import { getStockDetails } from "@/lib/stock/getStockDetails";
 import ProductCard from "@/components/admin/product/ProductCard"; 
-import { Product, ProductResponse } from "@/lib/admin/products/IProducts";
+import ProductFilters from "@/components/admin/product/ProductFilter";
+
+import { useProductList } from "@/hooks/product/useProductList";
 import {
   Pagination,
   PaginationContent,
@@ -15,124 +15,24 @@ import {
   PaginationPrevious,
   PaginationNext,
 } from "@/components/ui/pagination";
-import ProductFilters from "@/components/admin/product/ProductFilter";
-import useDebounce from "@/lib/admin/products/useDebounceHook";
 
 interface ProductListProps {
   token: string;
 }
-
 export default function ProductListPage({ token }: ProductListProps) {
   const router = useRouter();
 
-  // Estado de los filtros aplicados en la búsqueda
-  const [searchFilters, setSearchFilters] = useState({
-    code: "",
-    category: "",
-    minPrice: "",
-    maxPrice: "",
-    minCost: "",
-    maxCost: "",
-  });
+  const {
+    products,
+    stockMap,
+    isLoading,
+    pagination,
+    inputFilters,
+    setInputFilters,
+    handleSearch,
+    handlePageChange,
+  } = useProductList(token);
 
-  const [inputFilters, setInputFilters] = useState({ ...searchFilters });
-  
-  // Aplicamos debounce a los filtros de entrada (500ms de retraso)
-  const debouncedFilters = useDebounce(inputFilters, 500);
-  
-  const [products, setProducts] = useState<Product[]>([]);
-  const [stockMap, setStockMap] = useState<Record<string, number>>({});
-  const [isLoading, setIsLoading] = useState(false);
-  const [pagination, setPagination] = useState({
-    currentPage: 1,
-    totalPages: 1,
-    totalItems: 0,
-    pageSize: 5,
-  });
-
-  // Función para cargar el stock de un producto
-  const loadProductStock = useCallback(
-    async (productId: string) => {
-      try {
-        const stockData = await getStockDetails(productId, token);
-        const totalStock = stockData.data.reduce(
-          (total, detail) => total + detail.amount,
-          0
-        );
-        return totalStock;
-      } catch (error) {
-        console.error(`Error al cargar el stock del producto ${productId}:`, error);
-        return 0;
-      }
-    },
-    [token]
-  );
-
-  // La consulta se basa en searchFilters
-  const loadProducts = useCallback(
-    async (page: number, filterParams = searchFilters) => {
-      setIsLoading(true);
-      try {
-        const params = { ...filterParams, page, size: pagination.pageSize };
-        const data: ProductResponse = await getProducts(params, token);
-
-        // Cargar el stock para cada producto
-        const stockPromises = data.data.map(async (product) => {
-          const productStock = await loadProductStock(product.id);
-          return { id: product.id, stock: productStock };
-        });
-
-        const stockResults = await Promise.all(stockPromises);
-        
-        const newStockMap: Record<string, number> = {};
-        stockResults.forEach(result => {
-          newStockMap[result.id] = result.stock;
-        });
-        
-        setStockMap(newStockMap);
-        setProducts(data.data);
-        setPagination({
-          currentPage: page,
-          totalPages: data.totalPages,
-          totalItems: data.total,
-          pageSize: data.size,
-        });
-      } catch (error) {
-        console.error("Error al obtener productos", error);
-      } finally {
-        setIsLoading(false);
-      }
-    },
-    [searchFilters, token, pagination.pageSize, loadProductStock]
-  );
-
-  // Actualizamos los filtros con debounce
-  useEffect(() => {
-    if (token) {
-      setSearchFilters(debouncedFilters);
-      loadProducts(1, debouncedFilters);
-    }
-  }, [debouncedFilters, token, loadProducts]);
-
-  // Carga inicial
-  useEffect(() => {
-    if (token) {
-      loadProducts(1);
-    }
-  }, [token, loadProducts]);
-
-  const handleSearch = () => {
-    setSearchFilters(inputFilters);
-    setPagination((prev) => ({ ...prev, currentPage: 1 }));
-    loadProducts(1, inputFilters);
-  };
-
-  const handlePageChange = (page: number) => {
-    if (page < 1 || page > pagination.totalPages) return;
-    setPagination((prev) => ({ ...prev, currentPage: page }));
-    loadProducts(page, searchFilters);
-  };
-  
   const preventInvalidKeys = (e: React.KeyboardEvent<HTMLInputElement>) => {
     if (e.key === "-" || e.key === "e") e.preventDefault();
   };
@@ -181,11 +81,7 @@ export default function ProductListPage({ token }: ProductListProps) {
           <PaginationItem>
             <PaginationPrevious
               onClick={() => handlePageChange(pagination.currentPage - 1)}
-              className={
-                pagination.currentPage <= 1
-                  ? "pointer-events-none opacity-50"
-                  : ""
-              }
+              className={pagination.currentPage <= 1 ? "pointer-events-none opacity-50" : ""}
             />
           </PaginationItem>
           {Array.from({ length: pagination.totalPages }, (_, i) => i + 1).map((page) => (
@@ -201,11 +97,7 @@ export default function ProductListPage({ token }: ProductListProps) {
           <PaginationItem>
             <PaginationNext
               onClick={() => handlePageChange(pagination.currentPage + 1)}
-              className={
-                pagination.currentPage >= pagination.totalPages
-                  ? "pointer-events-none opacity-50"
-                  : ""
-              }
+              className={pagination.currentPage >= pagination.totalPages ? "pointer-events-none opacity-50" : ""}
             />
           </PaginationItem>
         </PaginationContent>
