+++ src/lib/stock/getStock.ts
@@ -1,4 +1,4 @@
-// lib/admin/stock/getStocks.ts
+
 import { STOCK_API } from "@/lib/urls";
 import { StockQueryParams, StockResponse } from "./IStock";
